% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_counts}
\alias{deg_counts}
\title{Data frame with frequencies (absolute and relative) of DEGs per contrast}
\format{
A data frame with the frequencies (absolute and relative) of
up- and down-regulated genes in each contrast. Relative frequencies
are calculated relative to the total number of genes in the count matrix
used for differential expression analysis.
}
\usage{
data(deg_counts)
}
\description{
This object was obtained with \code{get_deg_counts()} using the example
data set \strong{deg_list}.
}
\examples{
data(deg_counts)
}
\keyword{datasets}
