% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_chlamy}
\alias{go_chlamy}
\title{Data frame with GO terms annotated to each gene of Chlamydomonas reinhardtii}
\format{
A 2-column data frame with columns \strong{gene} (character, gene
ID), and \strong{GO} (character, name of GO term.)
}
\usage{
data(go_chlamy)
}
\description{
Data were obtained from Phytozome and processed so that each row contains
only one GO term (long format).
}
\examples{
data(go_chlamy)
}
\keyword{datasets}
