% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HybridExpress-package.R
\docType{package}
\name{HybridExpress-package}
\alias{HybridExpress}
\alias{HybridExpress-package}
\title{HybridExpress: Comparative analysis of RNA-seq data for hybrids and their progenitors}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

HybridExpress can be used to perform comparative transcriptomics analysis of hybrids (or allopolyploids) relative to their progenitor species. The package features functions to perform exploratory analyses of sample grouping, identify differentially expressed genes in hybrids relative to their progenitors, classify genes in expression categories (N = 12) and classes (N = 5), and perform functional analyses. We also provide users with graphical functions for the seamless creation of publication-ready figures that are commonly used in the literature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almeidasilvaf/HybridExpress}
  \item Report bugs at \url{https://support.bioconductor.org/tag/HybridExpress}
}

}
\author{
\strong{Maintainer}: Fabricio Almeida-Silva \email{fabricio_almeidasilva@hotmail.com} (\href{https://orcid.org/0000-0002-5314-2964}{ORCID})

Authors:
\itemize{
  \item Lucas Prost-Boxoen (\href{https://orcid.org/0000-0003-2779-9097}{ORCID})
  \item Yves Van de Peer \email{yves.vandepeer@psb.vib-ugent.be} (\href{https://orcid.org/0000-0003-4327-3730}{ORCID})
}

}
\keyword{internal}
