% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_expression_partitions}
\alias{plot_expression_partitions}
\title{Plot expression partitions}
\usage{
plot_expression_partitions(partition_table, group_by = "Group")
}
\arguments{
\item{partition_table}{A data frame with genes per expression partition
as returned by \code{expression_partitioning()}.}

\item{group_by}{Character indicating the name of the variable
in \strong{partition_table} to use to group genes. One of "Group" or
"Class". Default: "Group".}
}
\value{
A ggplot object with a plot showing genes in each expression
partition.
}
\description{
Plot expression partitions
}
\examples{
data(deg_list)
partition_table <- expression_partitioning(deg_list)
plot_expression_partitions(partition_table)
}
