% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_expression_triangle}
\alias{plot_expression_triangle}
\title{Plot a triangle of comparisons of DEG sets among generations}
\usage{
plot_expression_triangle(deg_counts)
}
\arguments{
\item{deg_counts}{Data frame with number of differentially expressed
genes per contrast as returned by \code{get_deg_counts}.}
}
\value{
A ggplot object with an expression triangle.
}
\description{
Plot a triangle of comparisons of DEG sets among generations
}
\details{
The expression triangle plot shows the number of differentially expressed
genes (DEGs) for each contrast. Numbers in the center of the lines (in bold)
indicate the total number of DEGs, while numbers near boxes indicate
the number of up-regulated genes in each generation of the triangle.
}
\examples{
data(deg_counts)
plot_expression_triangle(deg_counts)
}
