% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_chlamy}
\alias{se_chlamy}
\title{Expression data (in counts) for 3 Chlamydomonas lines (P1, P2, and F1)}
\format{
A \code{SummarizedExperiment} object with an assay (count) and
colData.
}
\usage{
data(se_chlamy)
}
\description{
Two lines (referred to as parent 1 and parent 2) with different ploidy
levels were crossed to generate an allopolyploid (F1).
}
\examples{
data(se_chlamy)
}
\keyword{datasets}
