% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_list}
\alias{deg_list}
\title{List of differentially expressed genes for all contrasts}
\format{
A list of data frames with gene-wise test statistics for
differentially expressed genes for each contrast. Contrasts are
"P1_vs_P2", "P1_vs_F1", "P2_vs_F1", and "midparent_vs_F1".
}
\usage{
data(deg_list)
}
\description{
This object was obtained with \code{get_deg_list()} using the example
data set \strong{se_chlamy}.
}
\examples{
data(deg_list)
}
\keyword{datasets}
