% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_expression_partitions}
\alias{plot_expression_partitions}
\title{Plot expression partitions}
\usage{
plot_expression_partitions(partition_table, group_by = "Group", palette = NULL)
}
\arguments{
\item{partition_table}{A data frame with genes per expression partition
as returned by \code{expression_partitioning()}.}

\item{group_by}{Character indicating the name of the variable
in \strong{partition_table} to use to group genes. One of "Group" or
"Class". Default: "Group".}

\item{palette}{Character vector with color names to be used for each level
of the variable specified in \strong{group_by}.
If \strong{group_by = "Group"}, this must be a vector of length 12.
If \strong{group_by = "Class"}, this must be a vector of length 5.
If NULL, a default color palette will be used.}
}
\value{
A ggplot object with a plot showing genes in each expression
partition.
}
\description{
Plot expression partitions
}
\examples{
data(deg_list)
partition_table <- expression_partitioning(deg_list)
plot_expression_partitions(partition_table)
}
